Remove extractTransformNames
Based on a few samples, it's no longer possible to get to the same set of
transform names as before (eg. Tone => Tonsprache in German).

Therefore it's not possible to make extractTransformNames backwards
compatible. I'll consider adding a new API to extract this information
on demand, if someone comes up with a use case :)