Do not accidentally override `n` operand when generating `e` and `c` operands
Recently support for the `e` and `c` operands has been added. Without
realizing, the `n` operand is accidentally overwritten when the operand
variables are initialized. The `n` variable is used as a temporary
accumulator but this varibale might be accessed later in the plural
rules, so it's incorrect to change its value.